Rename file to resolve clippy lint
Lint was "module inception"
